+++ rotkehlchen/externalapis/coinpaprika.py
@@ -102,10 +102,13 @@ KNOWN_TO_MISS_FROM_PAPRIKA = (
     'SQL',
     'SSD',
     'STMX',
+    'SUKU',
     'SWARM',  # Swarmcoin  https://coinmarketcap.com/currencies/swarm/
     'SWAP',
     'SUTER',
     'SYNC',
+    'TEND',
+    'TRADE',
     'TRXBEAR',  # The 3x TRX short
     'TRXBULL',  # The 3x TRX long
     'TNC',  # TNC token